jsonschema: Switch to map based encoding (#76)
Provides a public interface for #35, and switches the internal encoding
logic to rely on maps, as propsed in #51. This makes the implementation
simpler and more straight forward.

Compared to the poposed solution 1 in #35 that would rely on passing an
io.Writer to an encoder method, this solution will more easily handle a
corner-case where a custom validator writes zero bytes to the io.Writer.

The reduction in complexity does not come for free though. This looks to
be 5x slower for very tiny schemas, and aproximately 2x slower for more
complex schemas. As jsonschema is used as API documentation, and can be
easily cached, this performance loss should still be quite acceptable.