schema.Reference now validated as normal field.
Allows for schema.Reference to be validated in the same way all other
field validators are validated. To achive this, the resource.Index
will pass on a new schema.ReferenceChecker implementation when calling
Compile on each resource. To achive this, schema.Schema and each
schema.FieldValidator implementation that may hold a reference, has been
changed to implement a new interface schema.ReferenceCompiler in place
of the schema.Compiler interface.

This is a breaking change to the public API, but simple programs that
don't rely on calling Compile on Schema or Resource, are likely not
affected.