Split schema.Query to a dedicated query package + refactor
Isolate query types to a dedicated package as mixing schema and query
semantic is confusing. The query parsing and validation are now
performed in two distinct steps. It is thus now possible to parse a
query without the need of a schema.Validator. The Query type now has a
Validate method dedicated to this task. The schema.Q method is thus no
longer necessary and is replaced by query.MustParse when used in schema
for dependencies.