Migrate resource.Lookup into query.Query (breaking change) (#115)
* Rename query.Query to query.Predicate
* Refactor reference batch resolution and request coalesing
* Replace the ResourceResolver callback by a query.Resource interface
* Unify styles for query projection and predicate parsers
* Convert query package tests to table tests
* Fuzz-test query parsers and fix some discovered crashers
* Vendor some storers so all tests are always against a compat version
* Refactor of the resource.Lookup to be part of the query package
* Remove query predicate simplifier

Having pagination and projection in the resource package (Lookup type)
and the query in the schema/query package feels odd. If we want the
schema and the query package to be really idependant, all the tools to
perform a query should be in the query package.

This change moves the resource.Lookup type into the query.Query. The
previous query.Query type becomes query.Predicate and is now a field of
the query.Query type. The former resource.Lookup selector field becomes
the Projection field of the query.Query type and the pagination (offset
and limit) are now embedded into the query instead of being passed as
separate parameters to the Storer.

This change breaks compatibility with Storers. They all now have to
replace resource.Lookup, offset and limit parameters by this new
query.Query type.

One cool thing about this, is that the previously private connection
validator to handle connection in the resource package is now a public
validator in the schema package. The query.Projection Eval method is now
able to handle connections and references without depending on the
resource package. This will helps people wanting to use schema and
schema/query independantly from the rest of the project. This does also
make the code cleaner and easier to understand.

Also with this change, the query predicate simplifier has been removed. 
When a parsed predicate does only contain a single $and, we use to
simplify the query by removing the wrapping And type.

This creates issues when the end contains tests on the same field like
for instance:

   {$and: [{foo: {$ne: "bar"}}, {foo: {$ne: "baz"}}]}

Simplified this would result in:

   {foo: {$ne: "bar"}, foo: {$ne: "baz"}}

Which as JSON would fold the keys and only keep the last condition.

Issue #112